Fix: check name uniqueness without running on dataflow
